moved node creation in different file
