+++ apps/options-board/components/Board/postItFactory.js
@@ -3,12 +3,7 @@ import {
   encodeTransformAttribute,
   decodeTransformAttribute
 } from 'board/utils/svg'
-
-export const SVG_NS_URI = 'http://www.w3.org/2000/svg'
-
-const DRAGGABLE_CLASS = 'draggable'
-const HEIGHT = 76
-const WIDTH = 127
+import postItNodeFactory from './postItNodeFactory'
 
 const isMineEvent = (event, index) => {
   const indexAttribute = event.target.getAttribute('data-index')
@@ -19,29 +14,8 @@ const isMineEvent = (event, index) => {
   return parseInt(indexAttribute, 10) === index
 }
 
-const createNode = ({ x, y, color, index, attrs = {} }) => {
-  const group = document.createElementNS(SVG_NS_URI, 'g')
-
-  group.setAttribute('class', DRAGGABLE_CLASS)
-  group.setAttribute('transform', encodeTransformAttribute({ x, y }))
-
-  const postIt = document.createElementNS(SVG_NS_URI, 'rect')
-  postIt.setAttribute('fill', color)
-  postIt.setAttribute('height', HEIGHT)
-  postIt.setAttribute('width', WIDTH)
-  postIt.setAttribute('data-index', index)
-
-  Object.keys(attrs).forEach(attributeName => {
-    postIt.setAttribute(attributeName, attrs[attributeName])
-  })
-
-  group.appendChild(postIt)
-
-  return group
-}
-
 const factory = ({ parent, x, y, color, index, attrs = {} }) => {
-  const node = createNode({ x, y, color, index, attrs })
+  const node = postItNodeFactory({ x, y, color, index, attrs })
   let active = false
   let offset
